% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicho_cor_x.R
\name{dicho.cor.x}
\alias{dicho.cor.x}
\title{Dichotomize Data}
\usage{
dicho.cor.x(X)
}
\arguments{
\item{X}{A numeric vector or data frame. Categorical variables and missing data are unacceptable.}
}
\description{
Dichotomization of continuous variables is frequently criticized for its cost of losing information. In this work, we showed that, for variables of bimodal Gaussian mixtures, we can properly preserve more information than commonly used splits by applying our proposed algorithm to obtain optimized cut points.
}
\examples{
\dontrun{
  #simulation data
  n1=100;n2=100
  m1=0;m2=4;s1=2;s2=1
  X1=rnorm(n1,m1,s1)
  X2=rnorm(n2,m2,s2)
  X=c(X1,X2)

  dicho.cor.x(X)
}
}
